+++ src/kernel/ports/ports.c
@@ -70,7 +70,7 @@ void init_serial_port(uint32_t pa) {
  */
 void initialize_ports() {
     ebda = (void*)physical_to_virtual((((uintptr_t)*BDA_EBDA)>>4));
-    text_mode_video_memory = (uint16_t*)physical_to_virtual((uint64_t)(0xB8000));
+    text_mode_video_memory = (uint16_t*)physical_to_virtual(0xB8000);
 
     com_ports_count = 4;
     memset(com_ports_address, 0, sizeof(uint32_t)*com_ports_count);
